Cleaned up and removed unused properties
Removed some code that wasn't in use in BoardManger.cs
and adde the MouseManager.cs script (InteractionManager might be a more suitable name)
in preperation for some bigger code changes to try and seperate some code from the Tile.cs script
and modifyingp that Tile.cs script to derive from a new Script called Node
so i can more easily handle future interactions with the board more easily and also reducing the size of the code
that Tile needs to keep track of. At least that's the idea.